source_code/master/rpi/rest_api_app/requirements.txt
flask==1.1.1
PySerial==3.4
crcmod==1.7
psycopg2==2.8.3
apscheduler==3.6.1